+++ src/lib/util/dom.ts
@@ -1,0 +1,7 @@
+// The `tagName` is an `Element` property, but `contentEditable` is only `HTMLElement`s,
+// and the possible inputs are all `HTMLElement`, so this type seems correct.
+// If needed for things like `EventTarget`s, callers should check `el instanceof HTMLElement`.
+export const is_editable = (el: HTMLElement): boolean => {
+	const {tagName} = el;
+	return tagName === 'INPUT' || tagName === 'TEXTAREA' || el.contentEditable === 'true';
+};